refactor: mulitplat util into copa pkg
