+++ internal/copa/platform.go
@@ -1,6 +1,4 @@
-// Package multiplatform provides utilities to detect if Docker images support multiple platforms.
-// TODO: refactor to internal/copa/platform.go
-package multiplatform
+package copa
 
 import (
 	"context"
@@ -10,27 +8,6 @@ import (
 
 	"github.com/docker/docker/client"
 )
-
-// CopaSupportedPlatforms lists all platforms that Copa can patch
-// Based on Copa documentation: https://project-copacetic.github.io/copacetic/website/multiplatform-patching
-// TODO: mv to copa internal pkg
-var CopaSupportedPlatforms = []string{
-	"linux/amd64",
-	"linux/arm64",
-	"linux/arm/v7",
-	"linux/arm/v6",
-	"linux/386",
-	"linux/ppc64le",
-	"linux/s390x",
-	"linux/riscv64",
-}
-
-// ImageInfo contains information about an image's platform support and availability
-type ImageInfo struct {
-	IsMultiPlatform bool
-	IsLocal         bool
-	Platform        []string // Available platforms (e.g., ["linux/amd64", "linux/arm64"])
-}
 
 // GetImageInfo checks if the given image reference supports multiple platforms
 // and whether it's available locally or requires remote access.
