+++ internal/copa/platform_test.go
@@ -1,4 +1,4 @@
-package multiplatform
+package copa
 
 import (
 	"context"
